Fix for Vexo's GetMouseX() == GetScreenWidth() - 1 when it should be zero bug and adding a FileTypes option dialog in the editor
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@986 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/engine/win32/win32_main.cpp
@@ -199,14 +199,14 @@ LRESULT CALLBACK SphereWindowProc(HWND window, UINT msg, WPARAM wparam, LPARAM l
     case WM_MOUSEMOVE: {
       SetCursor(NULL);
 
-      int x = LOWORD(lparam);
+      int x = ((int)(short)LOWORD(lparam));
       if (x < 0) {
         x = 0;
       } else if (x > GetScreenWidth() - 1) {
         x = GetScreenWidth() - 1;
       }
 
-      int y = HIWORD(lparam);
+      int y = ((int)(short)HIWORD(lparam));
       if (y < 0) {
         y = 0;
       } else if (y > GetScreenHeight() - 1) {
